requirements.txt
google.generativeai
python-dotenv